Improve Error Handling and Update README
Actions:

+ Improved error handling when nil values passed in,
so that they return consistent errors.

+ Infer type: :ceremonial, when nil type value
passed in.

+ Added .gem files to gitignore

+ Updated Error section in README

+ Added specs for new logic and separated out the
existing spec into separate files for each method
call.

+ Set https://rubygems.org as push host.

+ Updated version

Motivation:

+ Want to ensure that nil values return the same
error as other invalid argument types, when passed
into a requisite value.

+ Existing spec was getting too long, and not passing
Rubocop.
