+++ lib/uk_county_locator/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module UkCountyLocator
-  VERSION = '0.1.0'
+  VERSION = '0.1.1'
 end
